added support for raw design vars, and fully tested usage of unnamed variables. TODO: test named variables, with new des types
